+++ __tests__/integration.test.tsx
@@ -1,0 +1,112 @@
+import { Server } from 'http'
+import { Browser } from 'puppeteer'
+
+import spawn from 'cross-spawn'
+import path from 'path'
+import fs from 'fs'
+import puppeteer from 'puppeteer'
+import handler from 'serve-handler'
+import http from 'http'
+import rmfr from 'rmfr'
+import * as cheerio from 'cheerio'
+
+jest.setTimeout(30000)
+
+describe('hydration', () => {
+  beforeAll(() => {
+    buildFixture('basic')
+  })
+
+  test('server rendered output', () => {
+    const result = readOutputFile('basic', 'index')
+    const $ = cheerio.load(result)
+
+    // server renders correctly
+    expect($('#__next').html()).toContain(`<h1>foo</h1><h1>Headline</h1>
+<p>hello <!-- -->jeff</p><button>Count: <!-- -->0</button>
+<p class=\"context\">Context value: \"<!-- -->foo<!-- -->\"</p>
+<p>Some <strong class=\"custom-strong\">markdown</strong> content</p>
+<div class=\"alert alert-warning g-type-body\"><p>Alert</p></div>
+<div>I am a dynamic component.</div>`)
+  })
+
+  test('rehydrates correctly in browser', async () => {
+    // hydrates correctly
+    let browser: Browser, server: Server
+    browser = await puppeteer.launch()
+    const page = await browser.newPage()
+    page.on('console', (msg) => console.log(msg.text()))
+    server = await serveStatic('basic')
+
+    await page.goto('http://localhost:1235')
+
+    // @ts-expect-error
+    await page.waitForFunction(() => Boolean(window.__NEXT_HYDRATED))
+
+    await page.waitForSelector('button')
+
+    // click the button
+    await page.click('button')
+
+    // wait for react to render
+    await page.waitForFunction(() => {
+      return document.querySelector('button')?.innerText === 'Count: 1'
+    })
+
+    // pull text for elements we're testing hydrate on
+    const contextElementText = await page.$eval(
+      '.context',
+      // @ts-expect-error -- el is typed as Element, but reasonable to assume it is an HTMLElement at this point
+      (el) => el.innerText
+    )
+    // @ts-expect-error -- el is typed as Element, but reasonable to assume it is an HTMLElement at this point
+    const buttonText = await page.$eval('button', (el) => el.innerText)
+
+    expect(buttonText).toEqual('Count: 1')
+    expect(contextElementText).toEqual('Context value: "bar"')
+
+    // close the browser and dev server
+    await browser.close()
+    await new Promise((resolve) => server.close(resolve))
+  })
+})
+
+afterAll(async () => {
+  await rmfr(path.join(__dirname, 'fixtures/basic/out'))
+  await rmfr(path.join(__dirname, 'fixtures/basic/.next'))
+})
+
+//
+// utility functions
+//
+
+function buildFixture(fixture: string) {
+  spawn.sync('next', ['build'], {
+    stdio: 'inherit',
+    cwd: path.join(__dirname, 'fixtures', fixture),
+    env: { ...process.env, NODE_ENV: undefined, __NEXT_TEST_MODE: 'true' },
+  })
+  spawn.sync('next', ['export'], {
+    stdio: 'inherit',
+    cwd: path.join(__dirname, 'fixtures', fixture),
+    env: { ...process.env, NODE_ENV: undefined, __NEXT_TEST_MODE: 'true' },
+  })
+}
+
+function readOutputFile(fixture: string, name: string) {
+  return fs.readFileSync(
+    path.join(__dirname, 'fixtures', fixture, 'out', `${name}.html`),
+    'utf8'
+  )
+}
+
+function serveStatic(fixture: string): Promise<Server> {
+  return new Promise((resolve) => {
+    const server = http.createServer((req, res) =>
+      handler(req, res, {
+        public: path.join(__dirname, 'fixtures', fixture, 'out'),
+      })
+    )
+    server.listen(1235, () => resolve(server))
+  })
+}
